requirements.txt
Django>=3.1.0,<3.2.0
djangorestframework>=3.11.0,<3.12.0
psycopg2
Pillow
django-jet
django-leaflet
django-geojson
jsonpickle
tweepy
geocoder
unicode
unidecode
feedparser
beautifulsoup4
httplib2
google-api-python-client
django-cors-headers
openrouteservice
python-dateutil
django-ipware
mock

flake8>=3.8.0,<3.9.0
